+++ jboss-hornetq-testsuite/src/test/java/org/jboss/qa/tools/JBMAdminOperationsEAP5.java
@@ -256,7 +256,7 @@ public class JBMAdminOperationsEAP5 implements JMSOperations {
     }
 
     @Override
-    public void addAddressSettings(String address, String addressFullPolicy, int maxSizeBytes, int redeliveryDelay, long redistributionDelay, long pageSizeBytes) {
+    public void addAddressSettings(String address, String addressFullPolicy, long maxSizeBytes, int redeliveryDelay, long redistributionDelay, long pageSizeBytes) {
 
         removeAddressSettings(address);
 
@@ -1567,7 +1567,7 @@ public class JBMAdminOperationsEAP5 implements JMSOperations {
     }
 
     @Override
-    public void addAddressSettings(String containerName, String address, String addressFullPolicy, int maxSizeBytes, int redeliveryDelay, long redistributionDelay, long pageSizeBytes) {
+    public void addAddressSettings(String containerName, String address, String addressFullPolicy, long maxSizeBytes, int redeliveryDelay, long redistributionDelay, long pageSizeBytes) {
         logger.info("This operation is not supported: " + getMethodName());
     }
