+++ jboss-hornetq-testsuite/src/test/java/org/jboss/qa/tools/JMSOperations.java
@@ -19,12 +19,12 @@ public interface JMSOperations {
      * @param addressFullPolicy   address full policy (PAGE, DROP or BLOCK)
      * @param maxSizeBytes        The max bytes size
      * @param redeliveryDelay     Defines how long to wait before attempting
-     *                            redelivery of a cancelled message
+*                            redelivery of a cancelled message
      * @param redistributionDelay Defines how long to wait when the last
-     *                            consumer is closed on a queue before redistributing any messages
+*                            consumer is closed on a queue before redistributing any messages
      * @param pageSizeBytes       The paging size
      */
-    void addAddressSettings(String address, String addressFullPolicy, int maxSizeBytes, int redeliveryDelay, long redistributionDelay, long pageSizeBytes);
+    void addAddressSettings(String address, String addressFullPolicy, long maxSizeBytes, int redeliveryDelay, long redistributionDelay, long pageSizeBytes);
 
     /**
      * Adds backup attribute.
@@ -1222,7 +1222,7 @@ public interface JMSOperations {
 
     void removeAddressSettings(String serverName, String address);
 
-    void addAddressSettings(String containerName, String address, String addressFullPolicy, int maxSizeBytes, int redeliveryDelay,
+    void addAddressSettings(String containerName, String address, String addressFullPolicy, long maxSizeBytes, int redeliveryDelay,
                             long redistributionDelay, long pageSizeBytes);
 
     /**
